+++ src/main/scala/scorex/waves/http/WavesApiRoute.scala
@@ -12,8 +12,9 @@ import scorex.api.http.{NegativeFee, NoBalance, _}
 import scorex.app.Application
 import scorex.crypto.encode.Base58
 import scorex.transaction.LagonakiTransaction.ValidationResult
+import scorex.transaction.state.wallet.Payment
 import scorex.waves.settings.WavesSettings
-import scorex.waves.transaction.{ExternalPayment, WavesTransactionModule}
+import scorex.waves.transaction.{ExternalPayment, SignedPayment, WavesTransactionModule}
 import scorex.waves.wallet.Wallet
 
 import scala.util.{Failure, Success, Try}
@@ -23,15 +24,77 @@ import scala.util.{Failure, Success, Try}
 case class WavesApiRoute(override val application: Application)(implicit val context: ActorRefFactory)
   extends ApiRoute with CommonTransactionApiFunctions {
 
+  lazy val wallet = application.wallet
+
   val suspendedSenders = application.settings.asInstanceOf[WavesSettings].suspendedSenders
 
   // TODO asInstanceOf
   implicit lazy val transactionModule: WavesTransactionModule = application.transactionModule.asInstanceOf[WavesTransactionModule]
 
   override lazy val route = pathPrefix("waves") {
-    payment ~ address ~ signPayment ~broadcastSignedPayment
+    externalPayment ~ address ~ signPayment ~ broadcastSignedPayment ~ payment
   }
 
+  // TODO: Should be moved to Scorex
+
+  @Path("/payment")
+  @ApiOperation(value = "Send payment from wallet",
+    notes = "Send payment from wallet to another wallet",
+    httpMethod = "POST",
+    produces = "application/json",
+    consumes = "application/json")
+  @ApiImplicitParams(Array(
+    new ApiImplicitParam(
+      name = "body",
+      value = "Json with data",
+      required = true,
+      paramType = "body",
+      dataType = "scorex.transaction.state.wallet.Payment",
+      defaultValue = "{\n\t\"amount\":400,\n\t\"fee\":1,\n\t\"sender\":\"senderId\",\n\t\"recipient\":\"recipientId\"\n}"
+    )
+  ))
+  @ApiResponses(Array(new ApiResponse(code = 200, message = "Json with response or error")))
+  def payment: Route = path("payment") {
+    entity(as[String]) { body =>
+      withAuth {
+        postJsonRoute {
+          walletNotExists(wallet).getOrElse {
+            Try(Json.parse(body)).map { js =>
+              js.validate[Payment] match {
+                case err: JsError =>
+                  WrongTransactionJson(err).json
+                case JsSuccess(payment: Payment, _) =>
+                  val txOpt = transactionModule.createPayment(payment, wallet)
+                  txOpt match {
+                    case Some(tx) =>
+                      tx.validate match {
+                        case ValidationResult.ValidateOke =>
+                          val signed = SignedPayment(tx.timestamp, tx.amount, tx.fee, tx.recipient.toString,
+                            Base58.encode(tx.sender.publicKey), tx.sender.address, Base58.encode(tx.signature))
+                          Json.toJson(signed)
+
+                        case ValidationResult.InvalidAddress =>
+                          InvalidAddress.json
+
+                        case ValidationResult.NegativeAmount =>
+                          NegativeAmount.json
+
+                        case ValidationResult.NegativeFee =>
+                          NegativeFee.json
+
+                        case ValidationResult.NoBalance =>
+                          NoBalance.json
+                      }
+                    case None =>
+                      InvalidSender.json
+                  }
+              }
+            }.getOrElse(WrongJson.json)
+          }
+        }
+      }
+    }
+  }
 
   @Path("/create-signed-payment")
   @ApiOperation(value = "Sign payment",
@@ -120,7 +183,7 @@ case class WavesApiRoute(override val application: Application)(implicit val con
     )
   ))
   @ApiResponses(Array(new ApiResponse(code = 200, message = "Json with response or error")))
-  def payment: Route = path("external-payment") {
+  def externalPayment: Route = path("external-payment") {
     withCors {
       entity(as[String]) { body =>
         val resp = Try(Json.parse(body)).map { js =>
@@ -154,8 +217,8 @@ case class WavesApiRoute(override val application: Application)(implicit val con
       value = "Json with data",
       required = true,
       paramType = "body",
-      dataType = "scorex.waves.transaction.ExternalPayment",
-      defaultValue = "{\n\t\"timestamp\": 0,\n\t\"amount\":400,\n\t\"fee\":1,\n\t\"senderPublicKey\":\"senderPubKey\",\n\t\"recipient\":\"recipientId\",\n\t\"signature\":\"sig\"\n}"
+      dataType = "scorex.waves.transaction.SignedPayment",
+      defaultValue = "{\n\t\"timestamp\": 0,\n\t\"amount\":400,\n\t\"fee\":1,\n\t\"senderPublicKey\":\"senderPubKey\",\n\t\"senderAddress\":\"senderAddress\",\n\t\"recipient\":\"recipientId\",\n\t\"signature\":\"sig\"\n}"
     )
   ))
   @ApiResponses(Array(new ApiResponse(code = 200, message = "Json with response or error")))
@@ -163,18 +226,15 @@ case class WavesApiRoute(override val application: Application)(implicit val con
     withCors {
       entity(as[String]) { body =>
         val resp = Try(Json.parse(body)).map { js =>
-          js.validate[ExternalPayment] match {
+          js.validate[SignedPayment] match {
             case err: JsError =>
               WrongJson.json
-            case JsSuccess(payment: ExternalPayment, _) =>
+            case JsSuccess(payment: SignedPayment, _) =>
               Base58.decode(payment.senderPublicKey) match {
                 case Success(senderPubKeyBytes) =>
                   val senderAddress = Account.fromPublicKey(senderPubKeyBytes)
-                  if (suspendedSenders.contains(senderAddress))
-                    InvalidSender.json
-                  else
-                    broadcastPayment(payment)
-
+                  if (suspendedSenders.contains(senderAddress)) InvalidSender.json
+                  else broadcastPayment(payment)
                 case Failure(e) => InvalidSender.json
               }
           }
@@ -185,6 +245,33 @@ case class WavesApiRoute(override val application: Application)(implicit val con
     }
   }
 
+  private def broadcastPayment(payment: SignedPayment) = {
+    transactionModule.broadcastPayment(payment) match {
+      case Left(tx) =>
+        if (!tx.signatureValid) InvalidSignature.json
+        else {
+          tx.validate match {
+            case ValidationResult.ValidateOke =>
+              tx.json
+
+            case ValidationResult.InvalidAddress =>
+              InvalidAddress.json
+
+            case ValidationResult.NegativeAmount =>
+              NegativeAmount.json
+
+            case ValidationResult.NegativeFee =>
+              NegativeFee.json
+          }
+        }
+      case Right(e) => e match {
+        case ValidationResult.NoBalance => NoBalance.json
+        case ValidationResult.InvalidAddress => InvalidAddress.json
+      }
+    }
+  }
+
+  @Deprecated
   private def broadcastPayment(payment: ExternalPayment) = {
     transactionModule.broadcastPayment(payment) match {
       case Left(tx) =>