+++ src/main/scala/scorex/waves/transaction/SignedPayment.scala
@@ -1,4 +1,4 @@
-package scorex.waves.http
+package scorex.waves.transaction
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json.{JsPath, Reads, Writes}
@@ -8,7 +8,7 @@ case class SignedPayment(timestamp: Long,
                          fee: Long,
                          recipient: String,
                          senderPublicKey: String,
-                         senderAddress: String,
+                         sender: String,
                          signature: String)
 
 object SignedPayment {
@@ -18,7 +18,17 @@ object SignedPayment {
       (JsPath \ "fee").write[Long] and
       (JsPath \ "recipient").write[String] and
       (JsPath \ "senderPublicKey").write[String] and
-      (JsPath \ "senderAddress").write[String] and
+      (JsPath \ "sender").write[String] and
       (JsPath \ "signature").write[String]
     ) (unlift(SignedPayment.unapply))
+
+  implicit val paymentReads: Reads[SignedPayment] = (
+    (JsPath \ "timestamp").read[Long] and
+      (JsPath \ "amount").read[Long] and
+      (JsPath \ "fee").read[Long] and
+      (JsPath \ "recipient").read[String] and
+      (JsPath \ "senderPublicKey").read[String] and
+      (JsPath \ "sender").read[String] and
+      (JsPath \ "signature").read[String]
+    ) (SignedPayment.apply _)
 }